refactor: Reorganize components into AP/SAT specific folders
- Create dedicated folder structure for better organization:
  - src/components/features/ap/ - AP specific components
  - src/components/features/sat/ - SAT specific components
  - src/components/features/shared/ - Shared components

- Move AP components to ap/ folder:
  - APExamInstructionPage.tsx → ap/exams/
  - APExamPage.tsx → ap/exams/
  - APResultsPage.tsx → ap/results/

- Move SAT components to sat/ folder:
  - MockExamPage.tsx → sat/exams/
  - SATSectionSelectPage.tsx → sat/exams/
  - SATDetailedResultsPage.tsx → sat/results/
  - SATResultsPage.tsx → sat/results/
  - All results components → sat/results/components/

- Move shared components to shared/ folder:
  - ScoreReportPage.tsx → shared/results/

- Update all import paths to match new structure:
  - App routes updated to use new component paths
  - Internal component imports updated
  - Relative imports adjusted for new folder hierarchy

Benefits:
- Clear separation of AP vs SAT functionality
- Better code organization and maintainability
- Easier to locate and modify exam-type specific features
- Scalable structure for future exam types
- Consistent folder naming conventions
