+++ src/app/exam/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useRouter } from "next/navigation";
-import { MockExamPage } from "@/components/features/exams/MockExamPage";
+import { MockExamPage } from "@/components/features/sat/exams/MockExamPage";
 import type { Subject } from "@/types";
 
 export default function ExamPage() {
