+++ src/app/sat-detailed-results/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useRouter } from "next/navigation";
-import { SATDetailedResultsPage } from "@/components/features/results/SATDetailedResultsPage";
+import { SATDetailedResultsPage } from "@/components/features/sat/results/SATDetailedResultsPage";
 import type { Subject } from "@/types";
 
 export default function SATDetailedResults() {
