+++ src/app/sat-section-select/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useRouter } from "next/navigation";
-import { SATSectionSelectPage } from "@/components/features/exams/SATSectionSelectPage";
+import { SATSectionSelectPage } from "@/components/features/sat/exams/SATSectionSelectPage";
 import type { Subject } from "@/types";
 
 export default function SATSectionSelect() {
