+++ src/components/features/sat/results/SATDetailedResultsPage.tsx
