+++ src/components/features/sat/results/components/SectionScoreCard.tsx
