+++ src/components/features/sat/results/components/TypeAnalysisChart.tsx
